requirements.txt
# MedVAE + Diffusion SR Pipeline Requirements
# Python 3.11+ required
#
# Installation instructions:
# For CPU-only: pip install -r requirements.txt
# For CUDA 12.1: pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu121
# For CUDA 11.8: pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu118
#
# Or use UV (recommended):
# uv pip install -r requirements.txt

# ============================================================================
# PyTorch (CUDA 12.1 compatible versions)
# ============================================================================
torch==2.5.0
torchvision==0.20.0
torchaudio==2.5.0

# ============================================================================
# Core Scientific Computing
# ============================================================================
numpy==1.26.4
scipy==1.13.1
pandas==2.2.3
scikit-learn==1.6.1
scikit-image==0.24.0

# ============================================================================
# Image Processing
# ============================================================================
Pillow==11.2.1
opencv-python==4.10.0.84
matplotlib==3.10.1
seaborn==0.13.2
imagecodecs==2024.12.30

# ============================================================================
# Medical Imaging
# ============================================================================
medvae==0.1.7
nibabel==5.3.2
SimpleITK==2.4.1
nilearn==0.11.1
dicom2nifti==2.5.1
monai==1.4.0

# ============================================================================
# Deep Learning - Training & Utilities
# ============================================================================
accelerate==1.7.0
torchmetrics==1.7.2
torch-fidelity==0.3.0
einops==0.8.1

# ============================================================================
# Diffusion Models
# ============================================================================
diffusers==0.33.1

# ============================================================================
# Transformers & CLIP
# ============================================================================
transformers==4.52.4
open_clip_torch==2.24.0
clip@git+https://github.com/openai/CLIP.git

# ============================================================================
# Data Processing & Augmentation
# ============================================================================
batchgenerators==0.25.1
batchgeneratorsv2==0.2.3

# ============================================================================
# Real-ESRGAN Dependencies
# ============================================================================
basicsr==1.4.2
facexlib==0.2.5
gfpgan==1.3.5

# ============================================================================
# Configuration & Experiment Management
# ============================================================================
omegaconf==2.3.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
yacs==0.1.8
wandb==0.14.0

# ============================================================================
# Utilities
# ============================================================================
tqdm==4.67.1
pyrootutils==1.0.4
rich==13.9.4
gdown==5.2.0
requests==2.32.3
polars==0.19.10
